# A_CONFIG_OK/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API de gestion des agendas et alertes",
        "title": "Config API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/agendas": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "agendas"
                ],
                "summary": "Lister les agendas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agendas"
                ],
                "summary": "Créer un agenda",
                "parameters": [
                    {
                        "description": "Agenda",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Agenda"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Agenda"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
        "/agendas/{id}": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "agendas"
                ],
                "summary": "Récupérer un agenda",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID agenda",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "agendas"
                ],
                "summary": "Mettre à jour un agenda",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID agenda",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Agenda (le champ id du body est ignoré, l'URL fait foi)",
                        "name": "agenda",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Agenda"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "agendas"
                ],
                "summary": "Supprimer un agenda",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID agenda",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
        "/alerts": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "alerts"
                ],
                "summary": "Lister les alertes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filtrer par agenda_id",
                        "name": "agenda_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "alerts"
                ],
                "summary": "Créer une alerte",
                "parameters": [
                    {
                        "description": "Alerte",
                        "name": "alert",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Alert"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
        "/alerts/{id}": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "alerts"
                ],
                "summary": "Récupérer une alerte",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID alerte",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "alerts"
                ],
                "summary": "Mettre à jour une alerte",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID alerte",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Alerte",
                        "name": "alert",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Alert"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "alerts"
                ],
                "summary": "Supprimer une alerte",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID alerte",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.APIError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "models.Agenda": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Alert": {
            "type": "object",
            "properties": {
                "agenda_id": {
                    "type": "string"
                },
                "condition": {
                    "description": "e.g., \"always\", \"room_change\"",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "target": {
                    "description": "e.g., email address",
                    "type": "string"
                }
            }
        }
    }
}